Refactering for View
